:recycle: Merge test_helper contents into main compat modules
- Move constants and methods from test_helper.rb into syntax.rb and bundle.rb
- Remove redundant FluentCompatBase:: prefix for inherited constants
- Delete now-empty test_helper.rb files
