+++ spec/support/compat/bundle.rb
@@ -1,12 +1,29 @@
 # frozen_string_literal: true
 
 require_relative "base"
+require_relative "bundle/ast_converter"
 
 # Compatibility testing support for fluent-bundle
 module FluentCompatBundle
   include FluentCompatBase
   extend FluentCompatBase
-end
 
-require_relative "bundle/ast_converter"
-require_relative "bundle/test_helper"
+  SYNTAX_FIXTURES = FLUENT_JS_ROOT / "fluent-syntax" / "test" / "fixtures_structure"
+  private_constant :SYNTAX_FIXTURES
+
+  BUNDLE_FIXTURES = FLUENT_JS_ROOT / "fluent-bundle" / "test" / "fixtures_structure"
+  private_constant :BUNDLE_FIXTURES
+
+  module_function def all_fixtures
+    collect_fixtures({json_dir: BUNDLE_FIXTURES, ftl_dir: SYNTAX_FIXTURES})
+  end
+
+  def parse_ftl(source)
+    parser = Foxtail::Bundle::Parser.new
+    parser.parse(source)
+  end
+
+  def convert_to_json(entries)
+    ASTConverter.to_json_format(entries)
+  end
+end
